fix DateTime literal parsing
